Auto-sync support for Eversan scoreboards
This can sync all data coming from the scoreboard, which for these units is
the score, time, and period. Scoreboard brightness data is also available
on the serial link but that doesn't seem very useful here :)

A bit of hardware is required to make this work, as the Eversan scoreboards
output 62.5kbps RS-485 with 9 data bits. See hw_interface/eversan.c for an
example of how this can be made more useful using an ATmega328P.